MarketAnalysis/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
tradingview-ta>=3.3.0
polygon-api-client>=1.0.0
requests>=2.26.0
python-dotenv>=0.19.0

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.0.0
black>=21.0.0
isort>=5.0.0
flake8>=3.9.0
mypy>=0.910

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0
myst-parser>=0.15.0